changed cfsheet.py, sheet.py, and utils.py so that the new lissom.ty now matches the c++ simulator.
The computation of response of each connection field is now moved into
the Projection class. The response is stored in temp_activation of each
Projection so that it does not have to be computed every time. There are
some simple optimizations to decrease the number of function calls as well.