fixed cfsheet.py and sheet.py so that the all the unit tests are passed.
The way that the weights in a connection field are initialized was changed
quite significantly in the previous version of cfsheet (1.11). This breaks one
test because the weights were assumed to be uniform in that version. Now the
code first checks if the initial weight distribution is uniform random. If yes,
then the code will follow the same flow as version 1.11. If no, the flow will
fall back to 1.10. The main difference between these two version is that 1.11
implements find the set of connection field boundary that matched the C++
simulator and should be correct. In 1.10, the size seems to be off by one
in some cases.

To fix the code so that all kinds of initial weight distributions will have the
correct connection field boundary may require some significant changes in
kernelfactory.py.