Attempting to organize functions that iterate over the CFs of a CFProjection and those that act on a single array.
(As part of the more general idea that there are e.g. output functions
which apply to a Projection.)

For output functions, we have the single-array (normal mathematical) output
functions, e.g.:
OutputFn
  Identity
  DivisiveSumNormalize
We also have the functions that iterate through a CFProjection's CFs,
applying an OutputFn to each:
CFProjectionOuputFn
  CFProjectionIdentityOutputFn(single_cf_fn=Identity())
  CFProjectionDivisiveSumNormalize(single_cf_fn=DivisiveSumNormalize())

OutputFns are in topo.outputfns.basic, and CFProjectionOutputFns are
in topo.outputfns.projfns. Other Projection-level output functions
would also go in the projfns module.

The same applies to LearningFns and
CFProjectionLearningFns. ResponseFunctions have been renamed
CFProjectionResponseFunctions where they apply to a CFProjection. The
reorganization of LearningFns and ResponseFns is not complete yet.

Along with this reorganization, CFProjectionOutputFnParameter (as
distinct from OutputFnParameter) has been introduced (and similarly
for LearningFns). Implementation of these is not complete yet; notably
the __init__.py files in the outputfns, learningfns, and responsefns
packages need to be cleaned up.

In addition to this, minor cleanups may also have occurred, including:
skipping calls to __init__() methods if all they do is call the
superclass' __init__()s; documentation strings.